service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.10.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-multipart==0.0.6
transformers==4.53.0
torch==2.8.0
sentencepiece==0.2.1
deep-translator==1.11.4
selenium==4.15.2
Pillow==10.4.0
pytesseract==0.3.13
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.23
minio==7.2.0
opencv-python==4.12.0.88
surya-ocr==0.15.4